move selectors
